refactor plotting functions to have different folders, less messy
