+++ rqpy/__init__.py
@@ -2,7 +2,7 @@ from ._globals import HAS_SCDMSPYTOOLS, HAS_TRIGSIM
 from . import core
 from .core import *
 from . import plotting
-from .plotting import hist, scatter, densityplot, passageplot
+from .plotting._core import *
 from . import process
 from . import io
 from . import sim
